fastapi_requirements.txt
fastapi
uvicorn
sqlmodel
Annotated
elasticsearch
python-dotenv
pandas
langchain
langchain-core
langchain-community
langchain-groq
qdrant-client
langchain-qdrant 
langchain-huggingface
torch